refactor: move case objects to case.py
