+++ foundation/scrollcase/requirements.txt
@@ -1,3 +1,4 @@
 -e .
 uv
 ipykernel
+nbconvert
