+++ foundation/scrollcase/src/scrollcase/__init__.py
@@ -1,104 +1,9 @@
-from build123d import *
-import meshlib.mrmeshpy as mm
-import tempfile
-import trimesh
-import numpy as np
-from . import mesh
-from . import cylinder
-from . import alignment
+import importlib
 
-__all__ = ["cylinder", "mesh", "alignment"]
+__all__ = [ "case", "mesh"]
 
 
 def __getattr__(name):
     if name in __all__:
         return importlib.import_module("." + name, __name__)
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
-
-
-def generic_mount_disc() -> Solid:
-    with BuildPart() as mount:
-        with BuildSketch() as sketch:
-            Circle(65)
-            Rectangle(20, 20, mode=Mode.SUBTRACT)
-        extrude(amount=10)
-    return mount
-
-
-def alignment_ring(radius: float):
-    """Generate an alignment ring.
-
-    Args:
-        radius: Minor radius
-
-    Returns:
-        Alignment ring
-    """
-    with BuildPart() as part:
-        Cylinder(radius + 2, 3)
-        Cylinder(radius + 3, 1)
-        Cylinder(radius, 3, mode=Mode.SUBTRACT)
-
-    return part
-
-
-def honeycomb_cylinder(
-    r: float, h: float, t: float, nr: float = 6, gap: float = 2
-) -> Compound:
-    """Build a hollow cylinder with honeycomb cutouts.
-
-    Args:
-        r: Cylinder minor radius.
-        h: Cylinder height.
-        t: Cylinder thickness.
-        nr: Number of radial cutouts. Defaults to 6.
-        gap: Gap between cutouts. Defaults to 1.
-
-    Returns:
-        Hhoneycomb cylinder.
-    """
-    # NOTE(akoen): This is not quite right since this assumes that the plane is flat,
-    # which it is not.
-    width = (4 * np.pi * r) / (3 * nr) - 2 * gap / np.sqrt(3)
-    spacing = 3 / 2 * width + np.sqrt(3) * gap
-
-    height = width * np.cos(np.pi / 6)
-
-    nz = int(np.ceil(2 * h / (height + gap))) + 1
-
-    all_hexes = []
-    for iz in range(nz):
-        start_angle = 360 / (2 * nr) * (iz % 2)
-        hex = (
-            (Plane.YZ * RegularPolygon(width / 2, 6, major_radius=True))
-            .face()
-            .translate((0, 0, iz * (height / 2 + gap / 2)))
-        )
-        hex = extrude(hex, 3, both=True)
-        hexes = PolarLocations(r, nr, start_angle=start_angle) * hex
-        all_hexes.extend(hexes)
-
-    all_hexes = [h for h in all_hexes]
-
-    cyl = Cylinder(r + t, h, align=(Align.CENTER, Align.CENTER, Align.MIN)) - Cylinder(
-        r, h, align=(Align.CENTER, Align.CENTER, Align.MIN)
-    )
-
-    honey_cyl = cyl - all_hexes
-
-    return honey_cyl
-
-
-def show_meshlib(mesh: mm.Mesh):
-    with tempfile.NamedTemporaryFile("w", suffix=".stl") as f:
-        mm.saveMesh(mesh, f.name)
-        tri_mesh = trimesh.load(f.name)
-        axis = trimesh.creation.axis(origin_size=10)
-        scene = trimesh.Scene([tri_mesh, axis])
-        return scene.show()
-
-
-def copy_meshlib(mesh: mm.Mesh):
-    copy = mm.Mesh()
-    copy.addPart(mesh)
-    return copy
