+++ foundation/scrollcase/src/scrollcase/alignment.py
@@ -1,4 +1,3 @@
-from warnings import deprecated
 import numpy as np
 from scipy.spatial import ConvexHull
 import logging
